I hate you all, and I hate shared code as well.
